Add result test helpers.
