+++ packages/rut/tests/Result.test.tsx
@@ -1,0 +1,7 @@
+import { runResultTestSuite } from '../src/testing/resultSuite';
+import { render, renderAndWait } from '../src/testing/helpers';
+
+runResultTestSuite({
+  render,
+  renderAndWait,
+});
